+++ Packs/GitHub/Integrations/GitHub/GitHub.py
@@ -1224,6 +1224,69 @@ def get_github_actions_usage():
     return_outputs(readable_output=human_readable, outputs=ec, raw_response=usage_result)
 
 
+def list_check_runs(owner_name, repository_name, run_id, commit_id):
+    url_suffix = None
+
+    if run_id:
+        url_suffix = f"/repos/{owner_name}/{repository_name}/check-runs/{run_id}"
+    elif commit_id:
+        url_suffix = f"/repos/{owner_name}/{repository_name}/commits/{commit_id}/check-runs"
+    else:
+        return_error("You have to specify either the check run id of the head commit refernce")
+
+    check_runs = http_request(method="GET", url_suffix=url_suffix)
+
+    return [r for r in check_runs.get('check_runs')]
+
+
+def get_github_get_check_run():
+    """ List github check runs.
+
+    """
+    command_args = demisto.args()
+    owner_name = command_args.get('owner', '')
+    repository_name = command_args.get('repository', '')
+    run_id = command_args.get('run_id', '')
+    commit_id = command_args.get('commit_id', '')
+
+    check_run_result = []
+
+    check_runs = list_check_runs(owner_name=owner_name, repository_name=repository_name, run_id=run_id, commit_id=commit_id)
+
+    for check_run in check_runs:
+            check_run_id = check_run.get('id', '')
+            check_external_id = check_run.get('external_id', '')
+            check_run_name = check_run.get('name', '')
+            check_run_app_name = check_run['app'].get('name', '')
+            check_run_pr = check_run['pull_requests']
+            check_run_status = check_run.get('status', '')
+            check_run_conclusion = check_run.get('conclusion', '')
+            check_run_started_at = check_run.get('started_at', '')
+            check_run_completed_at = check_run.get('completed_at', '')
+            check_run_output = check_run.get('output', '')
+
+            check_run_result.append({
+                'CheckRunID': check_run_id,
+                'CheckExternalID': check_external_id,
+                'CheckRunName': check_run_name,
+                'CheckRunAppName': check_run_app_name,
+                'CheckRunPR': check_run_pr,
+                'CheckRunStatus': check_run_status,
+                'CheckRunConclusion': check_run_conclusion,
+                'CheckRunStartedAt': check_run_started_at,
+                'CheckRunCompletedAt': check_run_completed_at,
+                'CheckRunOutPut': check_run_output
+            })
+
+    ec = {
+        'GitHub.CheckRuns(val.ID == obj.CheckRunID)': check_run_result
+    }
+    human_readable = tableToMarkdown('Check Run Details', check_run_result,
+                                     headerTransform=string_to_table_header)
+
+    return_outputs(readable_output=human_readable, outputs=ec, raw_response=check_run_result)
+
+
 def fetch_incidents_command():
     last_run = demisto.getLastRun()
     if last_run and 'start_time' in last_run:
@@ -1250,6 +1313,28 @@ def fetch_incidents_command():
             incidents.append(inc)
             if updated_at > last_time:
                 last_time = updated_at
+
+    if demisto.params().get('fetch_pull_requests'):
+        pr_list = http_request(method='GET',
+                               url_suffix=PULLS_SUFFIX,
+                               params={
+                                   'state': 'open',
+                                   'sort': 'created',
+                                   'per_page': 30,
+                                   'page': 1
+                               })
+        for pr in pr_list:
+            updated_at_str = pr.get('created_at')
+            updated_at = datetime.strptime(updated_at_str, '%Y-%m-%dT%H:%M:%SZ')
+            if updated_at > start_time:
+                inc = {
+                    'name': pr.get('url'),
+                    'occurred': updated_at_str,
+                    'rawJSON': json.dumps(pr)
+                }
+                incidents.append(inc)
+                if updated_at > last_time:
+                    last_time = updated_at
 
     demisto.setLastRun({'start_time': datetime.strftime(last_time, '%Y-%m-%dT%H:%M:%SZ')})
     demisto.incidents(incidents)
@@ -1285,6 +1370,8 @@ COMMANDS = {
     'GitHub-is-pr-merged': is_pr_merged_command,
     'GitHub-create-pull-request': create_pull_request_command,
     'Github-get-github-actions-usage': get_github_actions_usage,
+    'Github-get-check-run': get_github_get_check_run,
+
 }
 
 '''EXECUTION'''
