--- Packs/GitHub/Integrations/GitHub_Temp/GitHub_description.md
@@ -1,18 +1,0 @@
-
- To access the requested repository, you need to provide 
- the username, repository name, and API key as parameters for the
- integration.
- 
- To generate your personal access token, visit:
- https://github.com/settings/tokens
- 
- The integration supports the following workflows:
- - Create an issue
- - Update an issue
- - Close an issue
- - List all issues
- - Search for specific issues
- - Get the download count for each release in a repository
- 
- 
-
